Move plugin execution primitives into plugins service
New interface locations:

- `services/plugins/apiexec` owns the HTTP and GraphQL API execution helpers
  used by declarative providers and invocation error classification.
- `services/plugins/paraminterp` owns provider parameter interpolation shared by
  declarative providers and invocation routing.

Examples:

```go
result, err := apiexec.Do(ctx, httpClient, apiexec.Request{
    Method:  http.MethodGet,
    BaseURL: baseURL,
    Path:    "/accounts/{account_id}",
    Params:  params,
    Token:   token,
})
```

```go
path := paraminterp.Interpolate("/accounts/{account_id}", map[string]string{
    "account_id": accountID,
})
```

Existing API execution tests moved with the package, preserving the HTTP
transport, query encoding, content type, and pagination contracts. This removes
`services/*` imports of `internal/apiexec` and `internal/paraminterp`.
